[refactor] Refactor filterByFileIndex in FileStoreScan
